backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiosmtplib==3.0.1
email-validator==2.1.0
python-multipart==0.0.6
asyncio==3.4.3
httpx==0.25.2
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
passlib[bcrypt]==1.7.4
openai==1.6.1
